backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
python-dotenv
google-generativeai
